requirements.txt
apache-airflow==2.9.1
pyspark==3.5.1
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
python-dateutil==2.9.0
